backend/requirements.txt

fastapi
sqlalchemy
psycopg2-binary
boto3
google-generativeai
bcrypt
python-multipart
uvicorn
python-docx
PyPDF2
